estudando Java

+++ src/main/java/jdbc/Produto.java
@@ -2,21 +2,21 @@ package jdbc;
 
 public class Produto {
 	
-	private int codigo;
+	private Long codigo;
 	private String nome;
 	private double preco;
 	
 	
-	public Produto(int codigo, String nome, double preco) {
+	public Produto(Long codigo, String nome, double preco) {
 		super();
 		this.codigo = codigo;
 		this.nome = nome;
 		this.preco = preco;
 	}
-	public int getCodigo() {
+	public Long getCodigo() {
 		return codigo;
 	}
-	public void setCodigo(int codigo) {
+	public void setCodigo(Long codigo) {
 		this.codigo = codigo;
 	}
 	public String getNome() {
